Rename methods on MappedPayload.
Switched to a naming scheme for MappedPayload methods that explicitly contain …IfValid to make more obvious when the code handed into those methods actually gets executed. While this makes the processing chain a bit more verbose, it allows us to change ….alwaysMap(…) et al to simply ….map(…) and avoid the "always" as it's slightly misleading as the code handed into the method is also not executed in the absence case.