Optional integration with YAVI validation framework.
MappedPayloads now has a dedicated ….validate(BiConsumer<T, Errors>) method that is syntactic sugar for ….alwaysPeek(…). It allow the convenient invocation of any kind of validation infrastructure.

We now also ship a YaviValidator that is a wrapper for a YAVI validator type additionally implementing BiConsumer<T, Errors> so that instances of it can directly be handed into the MappedPayloads pipeline. Spring Boot auto-configuration is in play to automatically register a bean instance for that for each YAVI validator found in the application context.

```
@Configuration
class MyConfig {

  @Bean
  Validator<Sample> sampleValidator() {
    return ValidatorBuilder.of(Sample.class)
      .… // register constraints here
      .build();
  }
}

@RestController
@RequiredArgsConstructor
class MyController {

  private final YaviValidator<Sample> sampleValidator;

  @PostMapping("/post")
  HttpEntity<?> post(@RequestBody Sample sample, Errors errors) {

    return MappedPayloads.of(sample, errors)
      .validate(sampleValidator) // invoke validator
      .concludeIfValid(…);
  }
}
```