Add HTMX support for partials.
Added a module that provides an API to easily render Thymeleaf fragments in a way that is similar to Hotwire Turbo Streams by wrapping the rendered fragments into <div hx-swap-oob="…">…</div> elements.

Added a TodoMVC sample using that abstraction that's also a copy of the Hotwire one but adapting the templates to rather use the HTMX model.